requirements.txt
RPi.GPIO>=0.7.0
pygame>=2.0.0
luma.core>=2.3.1
luma.oled>=3.8.1
evdev>=1.4.0
pillow>=8.0.0
